+++ wxc_sdk/person_settings/available_numbers.py
@@ -4,7 +4,7 @@ from typing import Optional
 from wxc_sdk.api_child import ApiChild
 from wxc_sdk.base import ApiModel, enum_str
 from wxc_sdk.base import SafeEnum as Enum
-from wxc_sdk.common import NumberState, OwnerType, NumberOwner
+from wxc_sdk.common import NumberState, NumberOwner, IdAndName
 from wxc_sdk.person_settings.common import ApiSelector
 from wxc_sdk.rest import RestSession
 
@@ -32,6 +32,7 @@ class AvailableNumber(ApiModel):
     #: numbers (non-mobile, non toll-free numbers). If `true` the `phoneNumber` is a service number; otherwise, it is
     #: a standard number.
     is_service_number: Optional[bool] = None
+    location: Optional[IdAndName] = None
     owner: Optional[NumberOwner] = None